More refactoringzzz
